Extend FunDef with an optional StorageSpec. Refs #22.
Unlike FunDecln, which is defined as:

```
FunDecl (Maybe StorageSpec) Type Ident [Param]
```

a FunDef does not include a StorageSpec (optional or not):

```
FunDef Type Ident [Param] [Decln] [Stmt]
```

This prevents us from adding the keyword static to local functions, which is a
requirement for compliance with MISRA C 2012 in Copilot.

This commit modifies FunDef to have an optional StorageSpec as well.